refactor: streamline politician image components by consolidating exports and updating image properties for consistency across AU, CA, and GB
