+++ src/app/ca/politicians/person/[dtsiSlug]/twitter-image.tsx
@@ -1,12 +1,13 @@
-import {
-  alt,
-  contentType,
-  dynamic,
-  generateOgImage,
-  runtime,
-  size,
-} from '@/components/app/pagePoliticianDetails/common/generateOgImage'
+import { generateOgImage } from '@/components/app/pagePoliticianDetails/common/generateOgImage'
 
-export { alt, contentType, dynamic, runtime, size }
+export const dynamic = 'error'
+export const runtime = 'edge'
+export const alt = 'Image of politician and their stance on crypto'
+export const size = {
+  width: 1200,
+  height: 630,
+}
+
+export const contentType = 'image/png'
 
 export default generateOgImage
